+++ apps/profe/src/routes/player/+page.svelte
@@ -9,16 +9,60 @@
     let video = $state<LoadedVideo | null>(null);
     let steps = $state<PlayerStep[]>([
         {
-            timestamp: 35,
+            timestamp: 37,
             action: {
-                title: "Foo",
-                desc: "Bar",
+                title: "Check de ejemplo",
+                desc: "Los checks son sencillamente pausas para mostrar algo que el estudiante debería leer.",
                 type: "check",
+            },
+        },
+        {
+            timestamp: 45,
+            action: {
+                title: "Responde",
+                desc: "¿Qué crees que es?",
+                type: "freestanding",
+                hideAt: 52,
+            },
+        },
+        {
+            timestamp: 77,
+            action: {
+                title: "¿Cuál NO es parte de un estudio?",
+                desc: "Elige la opción correspondiente",
+                type: "choose",
+                options: [
+                    {
+                        label: "Población",
+                        correct: false,
+                        explanation: "La población sí es parte de un estudio.",
+                    },
+                    {
+                        label: "Muestra",
+                        correct: false,
+                        explanation: "La muestra sí es parte de un estudio.",
+                    },
+                    {
+                        label: "Juan Manuel Moreno Bonilla presidente de la Junta de Andalucía",
+                        correct: true,
+                        explanation:
+                            "Juan Manuel Moreno Bonilla presidente de la Junta de Andalucía NO es parte de un estudio estadístico. Es un personaje público que nada que ver.",
+                    },
+                    {
+                        label: "Variable",
+                        correct: false,
+                        explanation: "La variable sí es parte de un estudio.",
+                    },
+                ],
             },
         },
     ]);
     onMount(() => {
         const id = new URLSearchParams(window.location.search).get("v") ?? "";
+        if (id == "") {
+            loading = false;
+            return;
+        }
         const v = ZPVDs.find((v) => v.url.endsWith(id));
         if (!v) {
             loading = false;
@@ -38,5 +82,5 @@
         el video), esto es un fallo mío y deberías informarme para poder corregirlo.
     </p>
 {:else}
-    <Player intro={34} videoId={video.id} {steps} />
+    <Player intro={34} {video} {steps} />
 {/if}