+++ packages/types/types.ts
@@ -49,6 +49,7 @@ export type PlayerStep = {
         desc: string;
     } & (
         | { type: "check" }
+        | { type: "freestanding"; hideAt: number }
         | { type: "choose"; options: { label: string; correct: boolean; explanation: string }[] }
         | {
               type: "tof";